Squashed commit of the following:
commit 8bbb87eb5ca16654d02c6965897f91171f7e0f0f
Author: Patrick Gartung <gartung@fnal.gov>
Date:   Mon Mar 20 19:20:45 2023 +0100

    Revert "Link cmsRunJE to jemalloc-prof?"

    This reverts commit 00c53b33a1852b5454d5e4d3bbfa83c8375098f6.

commit 4dc5a4abc87aa6158edd9538b37b42c58e77c4e4
Author: Patrick Gartung <gartung@fnal.gov>
Date:   Mon Mar 20 19:19:48 2023 +0100

    add \<use name="jemalloc-prof" to JeProfService Buildfile.xml; add correct invocation to warning message; setenv for test

commit 021fe4f656660cb6e535edf2ec128ba9d2527337
Author: Patrick Gartung <gartung@fnal.gov>
Date:   Thu Mar 16 10:53:51 2023 +0100

    Code format

commit 00c53b33a1852b5454d5e4d3bbfa83c8375098f6
Author: Patrick Gartung <gartung@fnal.gov>
Date:   Thu Mar 16 10:07:53 2023 +0100

    Link cmsRunJE to jemalloc-prof?

commit 8b237f8a27ddf768ba8d8e921cce9b92f6ad3ca0
Author: Patrick Gartung <gartung@fnal.gov>
Date:   Wed Mar 15 11:42:55 2023 +0100

    include instructions for enabling jemalloc profining in warning message

commit bd7808dd5772be0eb092a7cebd603ec00e4b681d
Author: Patrick Gartung <gartung@fnal.gov>
Date:   Wed Mar 15 11:15:28 2023 +0100

    Use std::once_flag and std::call_once so warning only appears once.

commit 065a5230efc5d0707f42b7ca5da9d0770c4f74ba
Author: Patrick Gartung <gartung@fnal.gov>
Date:   Mon Mar 13 16:55:24 2023 +0100

    Add useful logging back to free function.

commit 10ba7df0db5846f6bc2ff60912f21999db10da41
Author: Patrick Gartung <gartung@fnal.gov>
Date:   Fri Mar 10 23:20:08 2023 +0100

    Code format

commit 82aafcbb4d1e4a9485d5a39b30289df46fb54dcc
Author: Patrick Gartung <gartung@fnal.gov>
Date:   Fri Mar 10 23:16:12 2023 +0100

    Use cms::jeprof::makeHeapDump in the service. Remove the check on have_jemalloc_and_prof which is in anonymous namespace and not accessible.

commit 7619dbc71bb39c7b7d645876542e5c9e35c67fd4
Author: Patrick Gartung <gartung@fnal.gov>
Date:   Fri Mar 3 18:59:33 2023 +0100

    Move jemalloc functions into anonymous namespace. Move makeHeapDump to cms::jeprof namespace. Require filename as argument/.

commit 2bb37f4f9bd8302115f1a14ce0df4382ca5a5d3a
Author: Patrick Gartung <gartung@fnal.gov>
Date:   Tue Feb 28 20:54:38 2023 +0100

    Code format

commit df20eb2f81f422aaeba240edf2ea650ecda4e699
Author: Patrick Gartung <gartung@fnal.gov>
Date:   Tue Feb 28 19:58:29 2023 +0100

    Change the function to makeHeapDump with no arguments.

commit ca242aecdab25ac1bb525e7aa83caead5e45f31f
Author: Patrick Gartung <gartung@fnal.gov>
Date:   Tue Feb 28 19:36:20 2023 +0100

    Set format if one is not set

commit 1b0736f1ea77829e2338c75478c233389ada225e
Author: Patrick Gartung <gartung@fnal.gov>
Date:   Tue Feb 28 19:23:16 2023 +0100

    PerfTools/JeProf: Make a library that allows a call to jemalloc profile heap dump function.
