+++ PerfTools/JeProf/BuildFile.xml
@@ -1,0 +1,5 @@
+<use name="sockets"/>
+<use name="FWCore/MessageLogger"/>
+<export>
+  <lib name="1"/>
+</export>
