+++ PerfTools/JeProf/interface/jeprof.h
@@ -1,0 +1,4 @@
+
+namespace cms::jeprof {
+  void makeHeapDump(const char *);
+}
