+++ PerfTools/JeProf/test/BuildFile.xml
@@ -1,0 +1,8 @@
+<bin name="jeprof_warn_t" file="jeprof_t.cc">
+  <use name="jemalloc"/>
+  <use name="PerfTools/JeProf"/>
+</bin>
+<bin name="jeprof_nowarn_t" file="jeprof_t.cc">
+  <use name="jemalloc-prof"/>
+  <use name="PerfTools/JeProf"/>
+</bin>
